Rename `main.rs` to `server.rs` in lobby subcrate
