+++ lobby/src/lib.rs
@@ -1,4 +1,4 @@
 #![feature(decl_macro)]
 #![feature(proc_macro_hygiene)]
 
-pub mod main;
+pub mod server;
